requirements.txt
absl-py==0.13.0
certifi==2021.5.30
colorama==0.4.4
contextlib2==21.6.0
cycler==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
ml-collections==0.1.0
numpy==1.21.2
Pillow==8.3.2
pynrrd==0.4.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
scipy==1.7.1
six==1.16.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.62.2
wincertstore==0.2
